pyproject.toml
[tool.poetry]
name = "language-table-dev"
version = "0.1.0"
description = ""
authors = ["nomutin <nomura0508@icloud.com>"]
readme = "README.md"
packages = [
	{ include = "language_table_dev" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"

[tool.poetry.group.core]
optional = true
[tool.poetry.group.core.dependencies]
tensorflow-datasets = "^4.9.2"
tensorflow-macos = "^2.13.0"
opencv-python = "^4.8.0.76"
matplotlib = "^3.7.2"
einops = "^0.6.1"
natsort = "^8.4.0"

[tool.poetry.group.lint]
optional = true
[tool.poetry.group.lint.dependencies]
ruff = "^0.0.284"
isort = "^5.12.0"
black = "^23.7.0"
mypy = "^1.5.0"

[tool.black]
line-length = 79

[tool.ruff]
select = ["ALL"]
ignore = ["D203", "D212"]
fixable = ["ALL"]
unfixable = []
exclude = [".git", ".ruff_cache", ".venv", "language-table"]
line-length = 79
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py310"

[tool.isort]
profile = "black"
line_length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
